README.md
Real-Time Camera Streaming and Recording with Flask is a Python project that demonstrates live camera streaming over a web interface using Flask. It captures video frames from the camera in real-time, adds date, time, and seconds to the video stream, and allows for automatic recording when faces or bodies are detected. The project offers adjustable frame rates for streaming and recording and is suitable for various real-time surveillance and monitoring applications.